Ladewertberechnung in FUNCTIONS/fun_Ladewert.py ausgelagert
