+++ FUNCTIONS/SymoGen24Connector.py
@@ -225,24 +225,6 @@ class SymoGen24:
         batterie_power_total = self.read_data('Battery_DC_Power_out') - self.read_data('Battery_DC_Power_in')
         batterie_power_scale_tmp = np.float64(np.int16(self.read_data('MPPT_Power_Scale_Factor')))
         return int(batterie_power_total * (10 ** batterie_power_scale_tmp))
-
-    def get_API(self):
-        import requests
-        import json
-        config = loadConfig('config.ini')
-        gen24url = "http://"+config['gen24']['hostNameOrIp']+"/components/readable"
-        url = requests.get(gen24url)
-        text = url.text
-        data = json.loads(text)
-        API = {}
-        API['AC_Produktion'] =  int(data['Body']['Data']['327680']['channels']['ACBRIDGE_ENERGYACTIVE_PRODUCED_SUM_U64']/3600)
-        API['DC_Produktion'] = int((data['Body']['Data']['393216']['channels']['PV_ENERGYACTIVE_ACTIVE_SUM_01_U64']+ data['Body']['Data']['393216']['channels']['PV_ENERGYACTIVE_ACTIVE_SUM_02_U64'])/3600)
-        API['Batterie_IN'] =    int(data['Body']['Data']['393216']['channels']['BAT_ENERGYACTIVE_ACTIVECHARGE_SUM_01_U64']/3600)
-        API['Batterie_OUT'] =   int(data['Body']['Data']['393216']['channels']['BAT_ENERGYACTIVE_ACTIVEDISCHARGE_SUM_01_U64']/3600)
-        API['Netzverbrauch'] =  int(data['Body']['Data']['16252928']['channels']['SMARTMETER_ENERGYACTIVE_CONSUMED_SUM_F64'])
-        API['Einspeisung'] =    int(data['Body']['Data']['16252928']['channels']['SMARTMETER_ENERGYACTIVE_PRODUCED_SUM_F64'])
-        return(API)
-
         
 # Test program
 if __name__ == "__main__":
