+++ html/3_tab_Hilfe.html
@@ -26,7 +26,7 @@
 <center>
 
 <font style="font-size: 24pt">
-<b>  GEN24_Ladesteuerung Version: 0.14.1 </b>
+<b>  GEN24_Ladesteuerung Version: 0.20.0 </b>
 </font>
 <hr>
 <br/><br/>
